requirement.txt
gradio>=5.0
groq
python-dotenv
openai-whisper
torch